Performing reconstruction of the codebase.
